+++ sdk/cosmosdb/cosmos/src/index.ts
@@ -60,6 +60,9 @@ export {
   TriggerOperation,
   TriggerType,
   UserDefinedFunctionType,
+  VectorEmbeddingPolicy,
+  VectorIndex,
+  VectorEmbedding,
 } from "./documents";
 
 export { UniqueKeyPolicy, UniqueKey } from "./client/Container/UniqueKeyPolicy";