+++ sdk/cosmosdb/cosmos/src/queryExecutionContext/index.ts
@@ -12,3 +12,4 @@ export * from "./parallelQueryExecutionContextBase";
 export * from "./parallelQueryExecutionContext";
 export * from "./orderByQueryExecutionContext";
 export * from "./pipelinedQueryExecutionContext";
+export * from "./orderByComparator";