+++ sdk/cosmosdb/cosmos/src/queryExecutionContext/pipelinedQueryExecutionContext.ts
@@ -2,7 +2,11 @@
 // Licensed under the MIT license.
 import { ClientContext } from "../ClientContext";
 import { Response, FeedOptions, QueryOperationOptions } from "../request";
-import { PartitionedQueryExecutionInfo } from "../request/ErrorResponse";
+import {
+  PartitionedQueryExecutionInfo,
+  QueryInfo,
+  // nonStreamingEndpointEmptyResult,
+} from "../request/ErrorResponse";
 import { CosmosHeaders } from "./CosmosHeaders";
 import { OffsetLimitEndpointComponent } from "./EndpointComponent/OffsetLimitEndpointComponent";
 import { OrderByEndpointComponent } from "./EndpointComponent/OrderByEndpointComponent";
@@ -17,7 +21,9 @@ import { GroupByValueEndpointComponent } from "./EndpointComponent/GroupByValueE
 import { SqlQuerySpec } from "./SqlQuerySpec";
 import { DiagnosticNodeInternal } from "../diagnostics/DiagnosticNodeInternal";
 import { RUCapPerOperationExceededErrorCode } from "../request/RUCapPerOperationExceededError";
-import { RUConsumedManager } from "../common";
+import { Constants, RUConsumedManager } from "../common";
+import { NonStreamingOrderByDistinctEndpointComponent } from "./EndpointComponent/NonStreamingOrderByDistinctEndpointComponent";
+import { NonStreamingOrderByEndpointComponent } from "./EndpointComponent/NonStreamingOrderByEndpointComponent";
 
 /** @hidden */
 export class PipelinedQueryExecutionContext implements ExecutionContext {
@@ -25,7 +31,11 @@ export class PipelinedQueryExecutionContext implements ExecutionContext {
   private fetchMoreRespHeaders: CosmosHeaders;
   private endpoint: ExecutionContext;
   private pageSize: number;
+  private vectorSearchBufferSize: number = 0;
   private static DEFAULT_PAGE_SIZE = 10;
+  private static DEFAULT_VECTOR_SEARCH_BUFFER_SIZE = 2000;
+  private nonStreamingOrderBy = false;
+
   constructor(
     private clientContext: ClientContext,
     private collectionLink: string,
@@ -38,67 +48,101 @@ export class PipelinedQueryExecutionContext implements ExecutionContext {
     if (this.pageSize === undefined) {
       this.pageSize = PipelinedQueryExecutionContext.DEFAULT_PAGE_SIZE;
     }
+    // Pick between Nonstreaming and streaming endpoints
+    this.nonStreamingOrderBy = partitionedQueryExecutionInfo.queryInfo.hasNonStreamingOrderBy;
 
     // Pick between parallel vs order by execution context
     const sortOrders = partitionedQueryExecutionInfo.queryInfo.orderBy;
-    if (Array.isArray(sortOrders) && sortOrders.length > 0) {
-      // Need to wrap orderby execution context in endpoint component, since the data is nested as a \
-      //      "payload" property.
-      this.endpoint = new OrderByEndpointComponent(
-        new OrderByQueryExecutionContext(
-          this.clientContext,
-          this.collectionLink,
-          this.query,
-          this.options,
-          this.partitionedQueryExecutionInfo,
-        ),
+    // TODO: Currently we don't get any field from backend to determine streaming queries
+    if (this.nonStreamingOrderBy) {
+      this.vectorSearchBufferSize = this.calculateVectorSearchBufferSize(
+        partitionedQueryExecutionInfo.queryInfo,
       );
-    } else {
-      this.endpoint = new ParallelQueryExecutionContext(
+
+      const distinctType = partitionedQueryExecutionInfo.queryInfo.distinctType;
+      const context: ExecutionContext = new ParallelQueryExecutionContext(
         this.clientContext,
         this.collectionLink,
         this.query,
         this.options,
         this.partitionedQueryExecutionInfo,
       );
-    }
-    if (
-      Object.keys(partitionedQueryExecutionInfo.queryInfo.groupByAliasToAggregateType).length > 0 ||
-      partitionedQueryExecutionInfo.queryInfo.aggregates.length > 0 ||
-      partitionedQueryExecutionInfo.queryInfo.groupByExpressions.length > 0
-    ) {
-      if (partitionedQueryExecutionInfo.queryInfo.hasSelectValue) {
-        this.endpoint = new GroupByValueEndpointComponent(
-          this.endpoint,
-          partitionedQueryExecutionInfo.queryInfo,
+
+      if (distinctType === "None") {
+        this.endpoint = new NonStreamingOrderByEndpointComponent(
+          context,
+          sortOrders,
+          this.vectorSearchBufferSize,
+          partitionedQueryExecutionInfo.queryInfo.offset,
         );
       } else {
-        this.endpoint = new GroupByEndpointComponent(
-          this.endpoint,
+        this.endpoint = new NonStreamingOrderByDistinctEndpointComponent(
+          context,
           partitionedQueryExecutionInfo.queryInfo,
+          this.vectorSearchBufferSize,
         );
       }
-    }
-    // If top then add that to the pipeline. TOP N is effectively OFFSET 0 LIMIT N
-    const top = partitionedQueryExecutionInfo.queryInfo.top;
-    if (typeof top === "number") {
-      this.endpoint = new OffsetLimitEndpointComponent(this.endpoint, 0, top);
-    }
+    } else {
+      if (Array.isArray(sortOrders) && sortOrders.length > 0) {
+        // Need to wrap orderby execution context in endpoint component, since the data is nested as a \
+        //      "payload" property.
+        this.endpoint = new OrderByEndpointComponent(
+          new OrderByQueryExecutionContext(
+            this.clientContext,
+            this.collectionLink,
+            this.query,
+            this.options,
+            this.partitionedQueryExecutionInfo,
+          ),
+        );
+      } else {
+        this.endpoint = new ParallelQueryExecutionContext(
+          this.clientContext,
+          this.collectionLink,
+          this.query,
+          this.options,
+          this.partitionedQueryExecutionInfo,
+        );
+      }
+      if (
+        Object.keys(partitionedQueryExecutionInfo.queryInfo.groupByAliasToAggregateType).length >
+          0 ||
+        partitionedQueryExecutionInfo.queryInfo.aggregates.length > 0 ||
+        partitionedQueryExecutionInfo.queryInfo.groupByExpressions.length > 0
+      ) {
+        if (partitionedQueryExecutionInfo.queryInfo.hasSelectValue) {
+          this.endpoint = new GroupByValueEndpointComponent(
+            this.endpoint,
+            partitionedQueryExecutionInfo.queryInfo,
+          );
+        } else {
+          this.endpoint = new GroupByEndpointComponent(
+            this.endpoint,
+            partitionedQueryExecutionInfo.queryInfo,
+          );
+        }
+      }
+      // If top then add that to the pipeline. TOP N is effectively OFFSET 0 LIMIT N
+      const top = partitionedQueryExecutionInfo.queryInfo.top;
+      if (typeof top === "number") {
+        this.endpoint = new OffsetLimitEndpointComponent(this.endpoint, 0, top);
+      }
 
-    // If offset+limit then add that to the pipeline
-    const limit = partitionedQueryExecutionInfo.queryInfo.limit;
-    const offset = partitionedQueryExecutionInfo.queryInfo.offset;
-    if (typeof limit === "number" && typeof offset === "number") {
-      this.endpoint = new OffsetLimitEndpointComponent(this.endpoint, offset, limit);
-    }
+      // If offset+limit then add that to the pipeline
+      const limit = partitionedQueryExecutionInfo.queryInfo.limit;
+      const offset = partitionedQueryExecutionInfo.queryInfo.offset;
+      if (typeof limit === "number" && typeof offset === "number") {
+        this.endpoint = new OffsetLimitEndpointComponent(this.endpoint, offset, limit);
+      }
 
-    // If distinct then add that to the pipeline
-    const distinctType = partitionedQueryExecutionInfo.queryInfo.distinctType;
-    if (distinctType === "Ordered") {
-      this.endpoint = new OrderedDistinctEndpointComponent(this.endpoint);
-    }
-    if (distinctType === "Unordered") {
-      this.endpoint = new UnorderedDistinctEndpointComponent(this.endpoint);
+      // If distinct then add that to the pipeline
+      const distinctType = partitionedQueryExecutionInfo.queryInfo.distinctType;
+      if (distinctType === "Ordered") {
+        this.endpoint = new OrderedDistinctEndpointComponent(this.endpoint);
+      }
+      if (distinctType === "Unordered") {
+        this.endpoint = new UnorderedDistinctEndpointComponent(this.endpoint);
+      }
     }
   }
 
@@ -127,7 +171,13 @@ export class PipelinedQueryExecutionContext implements ExecutionContext {
     } else {
       this.fetchBuffer = [];
       this.fetchMoreRespHeaders = getInitialHeader();
-      return this._fetchMoreImplementation(diagnosticNode, operationOptions, ruConsumedManager);
+      return this.nonStreamingOrderBy
+        ? this._nonStreamingFetchMoreImplementation(
+            diagnosticNode,
+            operationOptions,
+            ruConsumedManager,
+          )
+        : this._fetchMoreImplementation(diagnosticNode, operationOptions, ruConsumedManager);
     }
   }
 
@@ -157,7 +207,6 @@ export class PipelinedQueryExecutionContext implements ExecutionContext {
           return { result: temp, headers: this.fetchMoreRespHeaders };
         }
       } else {
-        // append the result
         this.fetchBuffer.push(item);
         if (this.fetchBuffer.length >= this.pageSize) {
           // fetched enough results
@@ -180,5 +229,80 @@ export class PipelinedQueryExecutionContext implements ExecutionContext {
         throw err;
       }
     }
+  }
+
+  private async _nonStreamingFetchMoreImplementation(
+    diagnosticNode: DiagnosticNodeInternal,
+    operationOptions?: QueryOperationOptions,
+    ruConsumedManager?: RUConsumedManager,
+  ): Promise<Response<any>> {
+    try {
+      const { result: item, headers } = await this.endpoint.nextItem(
+        diagnosticNode,
+        operationOptions,
+        ruConsumedManager,
+      );
+      mergeHeaders(this.fetchMoreRespHeaders, headers);
+      if (item === undefined) {
+        // no more results
+        if (this.fetchBuffer.length === 0) {
+          return {
+            result: undefined,
+            headers: this.fetchMoreRespHeaders,
+          };
+        } else {
+          // Just give what we have
+          const temp = this.fetchBuffer;
+          this.fetchBuffer = [];
+          return { result: temp, headers: this.fetchMoreRespHeaders };
+        }
+      } else {
+        // append the result
+        const ruConsumed = await ruConsumedManager.getRUConsumed();
+        // TODO:
+        const maxRUAllowed =
+          operationOptions && operationOptions.ruCapPerOperation
+            ? operationOptions.ruCapPerOperation
+            : Constants.NonStreamingQueryDefaultRUThreshold;
+        if (typeof item !== "object") {
+          this.fetchBuffer.push(item);
+        } else if (Object.keys(item).length !== 0) {
+          this.fetchBuffer.push(item);
+        }
+        if (this.fetchBuffer.length >= this.pageSize) {
+          // fetched enough results
+          const temp = this.fetchBuffer.slice(0, this.pageSize);
+          this.fetchBuffer = this.fetchBuffer.splice(this.pageSize);
+          return { result: temp, headers: this.fetchMoreRespHeaders };
+        } else if (ruConsumed * 2 < maxRUAllowed) {
+          // recursively fetch more only if we have more than 50% RUs left.
+          return this._nonStreamingFetchMoreImplementation(
+            diagnosticNode,
+            operationOptions,
+            ruConsumedManager,
+          );
+        } else {
+          return { result: [], headers: this.fetchMoreRespHeaders };
+        }
+      }
+    } catch (err: any) {
+      mergeHeaders(this.fetchMoreRespHeaders, err.headers);
+      err.headers = this.fetchMoreRespHeaders;
+      if (err.code === RUCapPerOperationExceededErrorCode && err.fetchedResults) {
+        err.fetchedResults.push(...this.fetchBuffer);
+      }
+      if (err) {
+        throw err;
+      }
+    }
+  }
+
+  private calculateVectorSearchBufferSize(queryInfo: QueryInfo): number {
+    if (queryInfo.top === 0 || queryInfo.limit === 0) return 0;
+    return queryInfo.top
+      ? queryInfo.top
+      : queryInfo.limit
+        ? queryInfo.offset + queryInfo.limit
+        : PipelinedQueryExecutionContext.DEFAULT_VECTOR_SEARCH_BUFFER_SIZE;
   }
 }